suggestions/requirements.txt
grpcio==1.70.0
grpcio-tools==1.70.0
protobuf==5.29.3
watchdog==6.0.0
requests==2.31.0
google-genai==1.3.0
python-dotenv==1.0.1
